added a function to recalculate index
